added logging

+++ install-masters.ps1
@@ -4,8 +4,17 @@
 setx /M PSModulePath "C:\Program Files\WindowsPowerShell\Modules;C:\Windows\system32\WindowsPowerShell\v1.0\Modules\;c:\masters\scripts\login"
 
 #clean up previous installation
-Unregister-ScheduledTask -TaskName "masterslogon" -Confirm:$false
-Remove-Item -Recurse -Force c:\masters
+if (get-scheduledtask | where { $_.TaskName -eq "masterslogon" })
+{
+	Unregister-ScheduledTask -TaskName "masterslogon" -Confirm:$false
+	write-host "removed scheduled task"
+}
+if (test-path "c:\masters")
+{
+	Remove-Item -Recurse -Force c:\masters
+	write-host "removed masters folder"
+}
+
 
 #create folder structure
 new-item -ItemType "directory" -Path "c:\masters"
